+++ src/js/message-service.js
@@ -1,6 +1,6 @@
 angular.module('message-center.service')
 
-.factory('MessageService', function() {
+.factory('MessageService', function () {
 
   var messageCenterPositions = {
     TopLeft: "top left",
@@ -10,7 +10,7 @@ angular.module('message-center.service')
     Centered: "centered",
   };
 
-  var MessageService = function(config) {
+  var MessageService = function (config) {
     this.config = {
       position: messageCenterPositions.TopRight,
       disabled: false,
@@ -22,9 +22,9 @@ angular.module('message-center.service')
     this.history = [];
     this.listeners = {
       broadcast: [],
+      danger: [],
       disable: [],
       info: [],
-      error: [],
       success: [],
       warning: []
     };
@@ -33,7 +33,7 @@ angular.module('message-center.service')
 
   MessageService.prototype.Position = messageCenterPositions;
 
-  MessageService.prototype.registerListener = function(topic, fn) {
+  MessageService.prototype.registerListener = function (topic, fn) {
     if (topic === 'warn') {
       topic = 'warning';
     }
@@ -43,23 +43,23 @@ angular.module('message-center.service')
     this.listeners[topic].push(fn);
   };
 
-  MessageService.prototype.configure = function(config) {
+  MessageService.prototype.configure = function (config) {
     angular.extend(this.config, config);
   };
 
-  MessageService.prototype.getHistory = function() {
+  MessageService.prototype.getHistory = function () {
     return this.history;
   };
 
-  MessageService.prototype.getDisabledHistory = function() {
+  MessageService.prototype.getDisabledHistory = function () {
     return this.disabledHistory;
   };
 
-  MessageService.prototype.clearHistory = function() {
+  MessageService.prototype.clearHistory = function () {
     this.history = [];
   };
 
-  MessageService.prototype.broadcast = function(msg, opts) {
+  MessageService.prototype.broadcast = function (msg, opts) {
     var msgObj = {
       message: msg,
       opts: opts
@@ -69,59 +69,28 @@ angular.module('message-center.service')
     } else {
       this.history.push(msgObj);
     }
-    this.listeners.broadcast.forEach(function(fn) {
+    this.listeners.broadcast.forEach(function (fn) {
       fn(msg, opts);
     });
   };
 
-  MessageService.prototype.danger = function(msg, opts) {
-    opts = opts || {};
-    opts.color = 'danger';
-    this.listeners.danger.forEach(function(fn) {
-      fn(msg, opts);
-    });
-    this.broadcast(msg, opts);
-  };
+  var MS = new MessageService();
 
-  MessageService.prototype.error = function(msg, opts) {
+  function convenienceHandler(type, msg, opts) {
     opts = opts || {};
-    opts.color = 'danger';
-    this.listeners.error.forEach(function(fn) {
-      fn(msg, opts);
-    });
-    this.broadcast(msg, opts);
-  };
-
-  MessageService.prototype.info = function(msg, opts) {
-    opts = opts || {};
-    opts.color = 'info';
-    this.listeners.info.forEach(function(fn) {
-      fn(msg, opts);
-    });
-    this.broadcast(msg, opts);
-  };
-
-  MessageService.prototype.success = function(msg, opts) {
-    opts = opts || {};
-    opts.color = 'success';
-    this.listeners.success.forEach(function(fn) {
-      fn(msg, opts);
-    });
-    this.broadcast(msg, opts);
-  };
-
-  function warning(msg, opts) {
-    opts = opts || {};
-    opts.color = 'warning';
-    this.listeners.warning.forEach(function(fn) {
+    opts.color = type;
+    this.listeners[type].forEach(function (fn) {
       fn(msg, opts);
     });
     this.broadcast(msg, opts);
   }
 
-  MessageService.prototype.warn = warning;
+  MS.danger = convenienceHandler.bind(MS, 'danger');
+  MS.error = convenienceHandler.bind(MS, 'danger');
+  MS.info = convenienceHandler.bind(MS, 'info');
+  MS.success = convenienceHandler.bind(MS, 'success');
+  MS.warn = convenienceHandler.bind(MS, 'warning');
+  MS.warning = convenienceHandler.bind(MS, 'warning');
 
-  MessageService.prototype.warning = warning;
-
-  return new MessageService();
+  return MS;
 });